fix(vision): drop unnecessary i32 cast in the MiniMax-M3 processor test
`expected_rows = (t * gh * gw) as i32` casts an already-i32 expression to i32, which trips `clippy::unnecessary_cast` under `-D warnings`. Pre-existing since the MiniMax-M3 processor landed; surfaced now because this branch touches `src/distributed/pipeline`, which triggers the pipeline-parallel-ci "Clippy on PP modules" step (the only CI job that lints this path). Drop the redundant cast.

+++ src/vision/processors/minimax_m3.rs
@@ -267,7 +267,7 @@ mod tests {
         mlxcel_core::eval(&pixel_values);
         let shape = mlxcel_core::array_shape(&pixel_values);
         let (t, gh, gw) = grid[0];
-        let expected_rows = (t * gh * gw) as i32;
+        let expected_rows = t * gh * gw;
         assert_eq!(shape, vec![expected_rows, 1176]);
     }
